requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
dj-database-url==0.5.0
Django==3.0.6
django-heroku==0.3.1
django-import-export==2.1.0
et-xmlfile==1.0.1
gunicorn==20.0.4
heroku==0.1.4
idna==2.9
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
psycopg2==2.8.5
python-dateutil==1.5
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
